Docker Image/requirements.txt
aws-cdk-lib==2.144.0
constructs>=10.0.0,<11.0.0
aws_ddk_core>=1.4.0,<2.0.0
cdk-ecr-deployment==3.0.74

